requirements.txt
streamlit==1.43.1
groq==0.18.0
python-dotenv==1.0.1
pillow==11.1.0
pandas==2.2.3
requests==2.32.3
altair==5.5.0